+++ lib/src/route_arguments.dart
@@ -1,5 +1,9 @@
 /// A helper class which can be injected to routes.
 class RouteArguments {
+  /// The raw path of the activated route.
+  /// For example /home/foo
+  final String path;
+
   /// The variables from the path.
   final Map<String, String> pathVariables;
 
@@ -7,5 +11,5 @@ class RouteArguments {
   final Map<String, String> queryParameters;
 
   /// RouteArguments constructor
-  RouteArguments(this.pathVariables, this.queryParameters);
+  RouteArguments(this.path, this.pathVariables, this.queryParameters);
 }