+++ test/route_arguments_test.dart
@@ -3,13 +3,15 @@ import 'package:test/test.dart';
 
 void main() {
   test('constructor empty', () {
-    var args = RouteArguments({}, {});
+    var args = RouteArguments('', {}, {});
+    expect(args.path.isEmpty, isTrue);
     expect(args.pathVariables.isEmpty, isTrue);
     expect(args.queryParameters.isEmpty, isTrue);
   });
 
   test('constructor not empty', () {
-    var args = RouteArguments({'foo': 'bar'}, {'bar': 'baz'});
+    var args = RouteArguments('/home', {'foo': 'bar'}, {'bar': 'baz'});
+    expect(args.path, equals('/home'));
     expect(args.pathVariables, equals({'foo': 'bar'}));
     expect(args.queryParameters, equals({'bar': 'baz'}));
   });