+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Features:
 - Added the `RouteProvider.tag` constant to reduce magic symbols.
 - Added a ServiceProvider extension for registering the explorator services.
 - Added a `GlobalKey<NavigatorState>` to the ServiceProvider.
-- Added A `Routing` class. This class simplifies the routing without a BuildContext.
+- Added a `Routing` class. This class simplifies the routing without a BuildContext.
+- Added the `path` to the `RouteArguments` class.
 
 ## 1.0.0